requirements.txt
Flask==2.2.2
Pillow==9.3.0
pipenv==2022.12.19
requests==2.28.1
requests-oauthlib==1.3.1
virtualenv==20.17.1
virtualenv-clone==0.5.7

